EventQueue takes n optionala queue
