+++ tests/integration/sidecars/publisher_queue_utils_tests.py
@@ -4,10 +4,9 @@ from importlib import reload
 
 import gevent.monkey
 
+from baseplate.lib.message_queue import create_queue
 from baseplate.lib.message_queue import MessageQueue
-from baseplate.lib.message_queue import PosixMessageQueue
 from baseplate.lib.message_queue import QueueType
-from baseplate.lib.message_queue import RemoteMessageQueue
 from baseplate.sidecars import publisher_queue_utils
 
 
@@ -23,12 +22,8 @@ class GeventPatchedTestCase(unittest.TestCase):
 
 
 class PublisherQueueUtilTests(GeventPatchedTestCase):
-    def test_posix_queue(self):
-        queue: MessageQueue = publisher_queue_utils.create_queue(QueueType.POSIX, "/test", 5, 1000)
-        assert isinstance(queue, PosixMessageQueue)
-
     def test_posix_queue_get_put(self):
-        queue: MessageQueue = publisher_queue_utils.create_queue(QueueType.POSIX, "/test", 5, 1000)
+        queue: MessageQueue = create_queue(QueueType.POSIX, "/test", 5, 1000)
 
         test_message = bytes("message", "utf-8")
         test_message_2 = bytes("2nd message", "utf-8")
@@ -39,16 +34,9 @@ class PublisherQueueUtilTests(GeventPatchedTestCase):
         output = queue.get()
         assert output == test_message_2
 
-    def test_in_memory_create_queue(self):
-        with publisher_queue_utils.start_queue_server(host="127.0.0.1", port=9090):
-            queue: MessageQueue = publisher_queue_utils.create_queue(
-                QueueType.IN_MEMORY, "/test", 5, 1000
-            )
-            assert isinstance(queue, RemoteMessageQueue)
-
     def test_in_memory_queue_get_put(self):
         with publisher_queue_utils.start_queue_server(host="127.0.0.1", port=9090):
-            queue: MessageQueue = publisher_queue_utils.create_queue(
+            queue: MessageQueue = create_queue(
                 QueueType.IN_MEMORY, "/test", 5, 1000
             )
 
@@ -63,7 +51,7 @@ class PublisherQueueUtilTests(GeventPatchedTestCase):
 
     def test_in_memory_queue_alternate_port(self):
         with publisher_queue_utils.start_queue_server(host="127.0.0.1", port=9091):
-            queue: MessageQueue = publisher_queue_utils.create_queue(
+            queue: MessageQueue = create_queue(
                 QueueType.IN_MEMORY, "/test", 5, 1000, port=9091
             )
